feat(scrolltracking): add initial multinode tracker, DAV-306
